protect course page
